refactor zephyr and jira components
