+++ src/jira/formatting.ts
@@ -1,0 +1,117 @@
+// Helper functions for formatting data for JIRA API
+
+// Helper function to format text content for JIRA API v3
+export function formatJiraContent(
+  content: string | undefined,
+  defaultText: string = "No content provided"
+) {
+  return content
+    ? {
+        type: "doc",
+        version: 1,
+        content: [
+          {
+            type: "paragraph",
+            content: [
+              {
+                type: "text",
+                text: content,
+              },
+            ],
+          },
+        ],
+      }
+    : {
+        type: "doc",
+        version: 1,
+        content: [
+          {
+            type: "paragraph",
+            content: [
+              {
+                type: "text",
+                text: defaultText,
+              },
+            ],
+          },
+        ],
+      };
+}
+
+// Helper function to format description for JIRA API v3
+export function formatDescription(description: string | undefined) {
+  return formatJiraContent(description, "No description provided");
+}
+
+// Helper function to format acceptance criteria for JIRA API v3
+export function formatAcceptanceCriteria(criteria: string | undefined) {
+  // Check if criteria is undefined or empty
+  if (!criteria) {
+    return {
+      type: "doc",
+      version: 1,
+      content: [
+        {
+          type: "paragraph",
+          content: [
+            {
+              type: "text",
+              text: "No acceptance criteria provided",
+            },
+          ],
+        },
+      ],
+    };
+  }
+
+  // Split criteria by newlines to handle bullet points properly
+  const lines = criteria.split("\n");
+  const content = [];
+
+  for (const line of lines) {
+    const trimmedLine = line.trim();
+
+    // Skip empty lines
+    if (!trimmedLine) continue;
+
+    // Check if line is a bullet point
+    if (trimmedLine.startsWith("-") || trimmedLine.startsWith("*")) {
+      content.push({
+        type: "bulletList",
+        content: [
+          {
+            type: "listItem",
+            content: [
+              {
+                type: "paragraph",
+                content: [
+                  {
+                    type: "text",
+                    text: trimmedLine.substring(1).trim(),
+                  },
+                ],
+              },
+            ],
+          },
+        ],
+      });
+    } else {
+      // Regular paragraph
+      content.push({
+        type: "paragraph",
+        content: [
+          {
+            type: "text",
+            text: trimmedLine,
+          },
+        ],
+      });
+    }
+  }
+
+  return {
+    type: "doc",
+    version: 1,
+    content: content,
+  };
+}
