+++ src/jira/index.ts
@@ -1,0 +1,5 @@
+// Re-export all Jira functionality
+export * from "./types.js";
+export * from "./api.js";
+export * from "./formatting.js";
+export * from "./tools.js";
