+++ src/utils.ts
@@ -1,0 +1,72 @@
+import fetch from "node-fetch";
+
+// Helper function to get the internal Jira ID and project ID from a ticket key
+export async function getJiraIssueId(
+  ticketKey: string,
+  auth: string
+): Promise<{
+  success: boolean;
+  id?: string;
+  projectId?: string;
+  errorMessage?: string;
+}> {
+  const jiraUrl = `https://${process.env.JIRA_HOST}/rest/api/3/issue/${ticketKey}`;
+
+  try {
+    const response = await fetch(jiraUrl, {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Basic ${auth}`,
+      },
+    });
+
+    const responseData = (await response.json()) as {
+      id?: string;
+      errorMessages?: string[];
+      fields?: {
+        project?: {
+          id?: string;
+        };
+      };
+    };
+
+    if (!response.ok) {
+      console.error("Error fetching ticket:", responseData);
+
+      let errorMessage = `Status: ${response.status} ${response.statusText}`;
+      if (responseData.errorMessages && responseData.errorMessages.length > 0) {
+        errorMessage = responseData.errorMessages.join(", ");
+      }
+
+      return { success: false, errorMessage };
+    }
+
+    if (!responseData.id) {
+      return {
+        success: false,
+        errorMessage: "No issue ID found in response",
+      };
+    }
+
+    // Extract project ID from the response
+    const projectId = responseData.fields?.project?.id;
+    if (!projectId) {
+      console.log("Warning: Project ID not found in response");
+    } else {
+      console.log(`Found project ID: ${projectId}`);
+    }
+
+    return {
+      success: true,
+      id: responseData.id,
+      projectId,
+    };
+  } catch (error) {
+    console.error("Exception fetching ticket ID:", error);
+    return {
+      success: false,
+      errorMessage: error instanceof Error ? error.message : String(error),
+    };
+  }
+}
