+++ src/zephyr/index.ts
@@ -1,0 +1,5 @@
+// Re-export all Zephyr functionality
+export * from "./types.js";
+export * from "./auth.js";
+export * from "./test-steps.js";
+export * from "./tools.js";
